% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{cargas_pca}
\alias{cargas_pca}
\title{Obtiene las cargas factoriales del PCA}
\usage{
cargas_pca(pca_obj)
}
\arguments{
\item{pca_obj}{Objeto resultante de la función prcomp()}
}
\value{
Una matriz con las cargas factoriales de cada variable en cada componente principal.
}
\description{
Obtiene las cargas factoriales del PCA
}
