% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{contrib_pca}
\alias{contrib_pca}
\title{Calcula la contribucion absoluta de las variables a los componentes principales}
\usage{
contrib_pca(pca_obj)
}
\arguments{
\item{pca_obj}{Objeto resultante de la función prcomp()}
}
\value{
Una matriz con los valores de contribución de cada variable a cada componente principal
}
\description{
Calcula la contribucion absoluta de las variables a los componentes principales
}
