% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{cos2_pca}
\alias{cos2_pca}
\title{Calcula la contribución relativa (coseno cuadrado) de las variables a los componentes principales}
\usage{
cos2_pca(pca_obj)
}
\arguments{
\item{pca_obj}{Objeto resultante de la funcion prcomp()}
}
\value{
Matriz con los cosenos cuadrados de cada variable en cada componente.
}
\description{
Calcula la contribución relativa (coseno cuadrado) de las variables a los componentes principales
}
