% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{eda_analysis}
\alias{eda_analysis}
\title{analisi exploratorio sobre el data set}
\usage{
eda_analysis(datos)
}
\arguments{
\item{datos}{base de datos para el proceso}
}
\description{
analisi exploratorio sobre el data set
}
