% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{generar_histogramas}
\alias{generar_histogramas}
\title{muestra histogramas de frecuencia sobre cada variable del data set}
\usage{
generar_histogramas(data)
}
\arguments{
\item{data}{base de datos la cual es necesaria para el proceso}
}
\description{
muestra histogramas de frecuencia sobre cada variable del data set
}
