% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{graficar_cargas_pca}
\alias{graficar_cargas_pca}
\title{Grafica las cargas factoriales del PCA (Biplot)}
\usage{
graficar_cargas_pca(pca_obj, comp_x = 1, comp_y = 2)
}
\arguments{
\item{pca_obj}{Objeto resultante de la funcion prcomp()}

\item{comp_x}{Numero del componente principal a graficar en el eje X (por defecto, PC1).}

\item{comp_y}{Numero del componente principal a graficar en el eje Y (por defecto, PC2).}
}
\value{
Un grafico que muestra las cargas factoriales de cada variable en los componentes seleccionados.
}
\description{
Grafica las cargas factoriales del PCA (Biplot)
}
