% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{matriz_correlacion}
\alias{matriz_correlacion}
\title{genera la matriz de correlaciones entre las variables numericas}
\usage{
matriz_correlacion(data)
}
\arguments{
\item{data}{base de datos la cual se quiere estudiar}
}
\description{
genera la matriz de correlaciones entre las variables numericas
}
