% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{total_eda}
\alias{total_eda}
\title{Análisis exploratorio completo del dataset}
\usage{
total_eda(data)
}
\arguments{
\item{data}{Base de datos a analizar}
}
\description{
Análisis exploratorio completo del dataset
}
